refactor: additional module
put request functions in separate module; timezone differences are now calculating from the total offset (instead of local and city offsets separately)
